Use the cloud token in service discovery credentials
Before, it was not possible to access cloud credentials given in the cloud block through service discovery. Things like private registry module sources in Terraform Cloud must use `terraform login` to successfully authenticate.

This change shims the credentials source with the cloud token if no creds exist for the cloud hostname. This keeps the existing behavior of preferring terraform login for module resolution, but also is a convenience to callers of backend.token(), private module developers, and future services provided by Terraform Cloud.